refactor - extract appointments list component, closes #8
